+++ host-agent/src/handlers/mod.rs
@@ -3,4 +3,5 @@ pub mod exec;
 pub mod files;
 pub mod hosts;
 pub mod network;
+pub mod sysinfo;
 //pub mod process;